requirements.txt
fastapi
uvicorn
python-multipart
faker
python-dotenv
